+++ rando/static/js/application/easing.js
@@ -1,0 +1,13 @@
+// Extraction of https://github.com/gdsmith/jquery.easing/blob/master/jquery.easing.1.3.js
+jQuery.easing['jswing'] = jQuery.easing['swing'];
+
+jQuery.extend( jQuery.easing,
+{
+    def: 'easeOutQuad',
+    easeInOutExpo: function (x, t, b, c, d) {
+        if (t==0) return b;
+        if (t==d) return b+c;
+        if ((t/=d/2) < 1) return c/2 * Math.pow(2, 10 * (t - 1)) + b;
+        return c/2 * (-Math.pow(2, -10 * --t) + 2) + b;
+    },
+});